Remove initialization task
This patch moves all initialization outside of a FreeRTOS task. Since
some of our initialization functions use osDelay() and some use
HAL_Delay(), a unified function (sysDelay()) was created that calls the
appropriate function based on the state of the FreeRTOS kernel.

The order of initialization was changed in order to load the flight
configuration as soon as possible and execute the required actions
(e.g., servo initialization).

The current is the following:
 1. Target-specific initialization
 2. Logging initialization
 3. Storage (flash chip) initialization
 4. Flight configuration load and setup
 5. Device initialization (sensors, ADC, battery monitor)
 6. Task creation

The patch also reverts target.cc files back to C.
