+++ boards/cats_rev1Pro/src/init/tasks.h
@@ -1,6 +1,6 @@
 /*
  * CATS Flight Software
- * Copyright (C) 2021 Control and Telemetry Systems
+ * Copyright (C) 2022 Control and Telemetry Systems
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -18,4 +18,4 @@
 
 #pragma once
 
-[[noreturn]] void task_init(void *argument);
+void init_tasks();
