+++ boards/cats_rev1Pro/src/util/task_util.c
@@ -1,0 +1,32 @@
+/*
+ * CATS Flight Software
+ * Copyright (C) 2022 Control and Telemetry Systems
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+
+#include "util/task_util.h"
+#include "target.h"
+
+#include <stdbool.h>
+
+volatile bool rtos_started = false;
+
+void sysDelay(uint32_t delay) {
+  if (rtos_started) {
+    osDelay(delay);
+  } else {
+    HAL_Delay(delay);
+  }
+}
